Refactor eeX plus.security and plus.annotation classes to core
