+++ jetty-core/jetty-plus/src/main/java/module-info.java
@@ -13,11 +13,15 @@
 
 module org.eclipse.jetty.plus
 {
+    requires org.eclipse.jetty.security;
     requires org.eclipse.jetty.util;
     requires org.slf4j;
 
     requires transitive java.naming;
-    requires transitive java.sql;
+    // Only required if using DataSourceLoginService.
+    requires static java.sql;
 
+    exports org.eclipse.jetty.plus.annotation;
     exports org.eclipse.jetty.plus.jndi;
+    exports org.eclipse.jetty.plus.security;
 }
