+++ jetty-core/jetty-plus/src/main/java/org/eclipse/jetty/plus/annotation/Injection.java
@@ -11,7 +11,7 @@
 // ========================================================================
 //
 
-package org.eclipse.jetty.ee9.plus.annotation;
+package org.eclipse.jetty.plus.annotation;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.Member;
