+++ jetty-core/jetty-plus/src/main/java/org/eclipse/jetty/plus/annotation/InjectionCollection.java
@@ -11,7 +11,7 @@
 // ========================================================================
 //
 
-package org.eclipse.jetty.ee9.plus.annotation;
+package org.eclipse.jetty.plus.annotation;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
