+++ jetty-core/jetty-plus/src/main/java/org/eclipse/jetty/plus/annotation/PreDestroyCallback.java
@@ -11,7 +11,7 @@
 // ========================================================================
 //
 
-package org.eclipse.jetty.ee10.plus.annotation;
+package org.eclipse.jetty.plus.annotation;
 
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
