+++ jetty-core/jetty-plus/src/main/java/org/eclipse/jetty/plus/annotation/package-info.java
@@ -14,4 +14,4 @@
 /**
  * Jetty Plus : Limited JEE Annotation Support
  */
-package org.eclipse.jetty.ee9.plus.annotation;
+package org.eclipse.jetty.plus.annotation;
