+++ jetty-ee10/jetty-ee10-annotations/src/test/java/org/eclipse/jetty/ee10/annotations/resources/TestResourceAnnotations.java
@@ -21,9 +21,9 @@ import javax.naming.InitialContext;
 import org.eclipse.jetty.ee10.annotations.AnnotationIntrospector;
 import org.eclipse.jetty.ee10.annotations.ResourceAnnotationHandler;
 import org.eclipse.jetty.ee10.annotations.ResourcesAnnotationHandler;
-import org.eclipse.jetty.ee10.plus.annotation.Injection;
-import org.eclipse.jetty.ee10.plus.annotation.InjectionCollection;
 import org.eclipse.jetty.ee10.webapp.WebAppContext;
+import org.eclipse.jetty.plus.annotation.Injection;
+import org.eclipse.jetty.plus.annotation.InjectionCollection;
 import org.eclipse.jetty.plus.jndi.EnvEntry;
 import org.eclipse.jetty.server.Server;
 import org.junit.jupiter.api.AfterEach;
