+++ jetty-ee10/jetty-ee10-plus/src/main/java/module-info.java
@@ -21,16 +21,10 @@ module org.eclipse.jetty.ee10.plus
 
     requires transitive org.eclipse.jetty.ee10.webapp;
 
-    // Only required if using DataSourceLoginService.
-    requires static java.sql;
     // Only required if using Transaction.
     requires static jakarta.transaction;
-    // Only required if using RunAs.
-    requires static org.eclipse.jetty.ee10.servlet;
 
-    exports org.eclipse.jetty.ee10.plus.annotation;
     exports org.eclipse.jetty.ee10.plus.jndi;
-    exports org.eclipse.jetty.ee10.plus.security;
     exports org.eclipse.jetty.ee10.plus.webapp;
 
     provides org.eclipse.jetty.ee10.webapp.Configuration with
