--- jetty-ee10/jetty-ee10-plus/src/main/java/org/eclipse/jetty/ee10/plus/security/package-info.java
@@ -1,17 +1,0 @@
-//
-// ========================================================================
-// Copyright (c) 1995 Mort Bay Consulting Pty Ltd and others.
-//
-// This program and the accompanying materials are made available under the
-// terms of the Eclipse Public License v. 2.0 which is available at
-// https://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
-// which is available at https://www.apache.org/licenses/LICENSE-2.0.
-//
-// SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
-// ========================================================================
-//
-
-/**
- * Jetty Plus : Limited JEE Security Support
- */
-package org.eclipse.jetty.ee10.plus.security;
