+++ jetty-ee10/jetty-ee10-plus/src/main/java/org/eclipse/jetty/ee10/plus/webapp/PlusDecorator.java
@@ -13,9 +13,9 @@
 
 package org.eclipse.jetty.ee10.plus.webapp;
 
-import org.eclipse.jetty.ee10.plus.annotation.InjectionCollection;
-import org.eclipse.jetty.ee10.plus.annotation.LifeCycleCallbackCollection;
 import org.eclipse.jetty.ee10.webapp.WebAppContext;
+import org.eclipse.jetty.plus.annotation.InjectionCollection;
+import org.eclipse.jetty.plus.annotation.LifeCycleCallbackCollection;
 import org.eclipse.jetty.util.Decorator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
