+++ jetty-ee9/jetty-ee9-annotations/src/main/java/org/eclipse/jetty/ee9/annotations/PostConstructAnnotationHandler.java
@@ -18,11 +18,11 @@ import java.lang.reflect.Modifier;
 
 import jakarta.annotation.PostConstruct;
 import org.eclipse.jetty.ee9.annotations.AnnotationIntrospector.AbstractIntrospectableAnnotationHandler;
-import org.eclipse.jetty.ee9.plus.annotation.LifeCycleCallbackCollection;
-import org.eclipse.jetty.ee9.plus.annotation.PostConstructCallback;
 import org.eclipse.jetty.ee9.webapp.MetaData;
 import org.eclipse.jetty.ee9.webapp.Origin;
 import org.eclipse.jetty.ee9.webapp.WebAppContext;
+import org.eclipse.jetty.plus.annotation.LifeCycleCallbackCollection;
+import org.eclipse.jetty.plus.annotation.PostConstructCallback;
 
 public class PostConstructAnnotationHandler extends AbstractIntrospectableAnnotationHandler
 {
