+++ jetty-ee9/jetty-ee9-annotations/src/test/java/org/eclipse/jetty/ee9/annotations/resources/TestResourceAnnotations.java
@@ -21,9 +21,9 @@ import javax.naming.InitialContext;
 import org.eclipse.jetty.ee9.annotations.AnnotationIntrospector;
 import org.eclipse.jetty.ee9.annotations.ResourceAnnotationHandler;
 import org.eclipse.jetty.ee9.annotations.ResourcesAnnotationHandler;
-import org.eclipse.jetty.ee9.plus.annotation.Injection;
-import org.eclipse.jetty.ee9.plus.annotation.InjectionCollection;
 import org.eclipse.jetty.ee9.webapp.WebAppContext;
+import org.eclipse.jetty.plus.annotation.Injection;
+import org.eclipse.jetty.plus.annotation.InjectionCollection;
 import org.eclipse.jetty.plus.jndi.EnvEntry;
 import org.eclipse.jetty.server.Server;
 import org.junit.jupiter.api.AfterEach;
