+++ jetty-ee9/jetty-ee9-plus/src/main/java/module-info.java
@@ -22,16 +22,10 @@ module org.eclipse.jetty.ee9.plus
     requires transitive org.eclipse.jetty.ee9.webapp;
     requires transitive org.eclipse.jetty.plus;
 
-    // Only required if using DataSourceLoginService.
-    requires static java.sql;
     // Only required if using Transaction.
     requires static jakarta.transaction;
-    // Only required if using RunAs.
-    requires static org.eclipse.jetty.ee9.servlet;
 
-    exports org.eclipse.jetty.ee9.plus.annotation;
     exports org.eclipse.jetty.ee9.plus.jndi;
-    exports org.eclipse.jetty.ee9.plus.security;
     exports org.eclipse.jetty.ee9.plus.webapp;
 
     provides org.eclipse.jetty.ee9.webapp.Configuration with
