+++ jetty-ee9/jetty-ee9-plus/src/test/java/org/eclipse/jetty/ee9/plus/annotation/LifeCycleCallbackCollectionTest.java
@@ -19,6 +19,10 @@ import java.nio.file.Path;
 import jakarta.servlet.http.HttpServlet;
 import org.eclipse.jetty.ee9.servlet.ServletHolder;
 import org.eclipse.jetty.ee9.webapp.WebAppContext;
+import org.eclipse.jetty.plus.annotation.LifeCycleCallback;
+import org.eclipse.jetty.plus.annotation.LifeCycleCallbackCollection;
+import org.eclipse.jetty.plus.annotation.PostConstructCallback;
+import org.eclipse.jetty.plus.annotation.PreDestroyCallback;
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.toolchain.test.FS;
 import org.eclipse.jetty.toolchain.test.jupiter.WorkDir;
